requirements.txt
fastapi
uvicorn
python-multipart
google-cloud-vision
openai
python-dotenv